requirements.txt
customtkinter
requests
pillow
packaging
